Initial commit.
The code works but is ugly. Horribly ugly.

There are two encoding formats in use here: a simple unsigned version,
and a more complicated signed version.

The unsigned version operates on a simple principle: the top 4 bits are
the number of bytes long that the value needs for representation,
allowing for the header to also stuff up to 4 bytes in it. This allows
encoding unsigned integers with 124 bits of information using a variable
length of 1 byte through 16 bytes, all which can be memcmp'ed to sort
properly.

The signed version uses 5 bits to encode a signed length in the first
byte. The complications arise from needing to manually handle the 2's
compliment for the, and needing to flip all the "truncated" bits when
the sign is negative.

The result is again a variable length of 1 to 16 bytes which can be
memcmp'ed to sort. The effective range is -2^123..-2^123 (using Rust
notation).