Added implementations for isize and usize
Originally I steered away from these due to the size varying on
different platforms being not a great design from a database
perspective. However, on reflection, variable integer encoding of
isize/usize makes them able to be parsed correctly on different
architectures. So, if a user has data encoded for usize that is >
u32::MAX, and they parse the data into a usize on a 32 bit platform,
they will receive an error. But, this support allows the developer to
change the key type from usize to u64, and the decoding will still work
correctly.