Improving code coverage
While trying to refactor the decode algorithms to work with all byte
sizes, there was an edge case before I refactored to a match statement
where N had to be greater than 1 for that block to execute. However, the
match statement moved the 0/1 cases to the top, preventing that
condition from ever being required.

Additionally, due to direct encoding being done, some of the try_from
paths for Signed/Unsigned weren't being tested.